@@ -1,11 +1,15 @@
-﻿namespace HeimerNET.Lor.DataDragon
+﻿using System.Text.Json.Serialization;
+
+namespace HeimerNET.Lor.DataDragon;
+
+public record SpellSpeed
 {
-    public record SpellSpeed
-    {
-        /// <summary>
-        /// Localized <see cref="NameRef"/>.
-        /// </summary>
-        public string Name { get; init; } = default!;
-        public string NameRef { get; init; } = default!;
-    }
+    /// <summary>
+    /// Localized <see cref="NameRef"/>.
+    /// </summary>
+    [JsonPropertyName("name")]
+    public string Name { get; init; } = default!;
+
+    [JsonPropertyName("nameRef")]
+    public string NameRef { get; init; } = default!;
 }
