@@ -1,22 +1,21 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace HeimerNET.Lor.GameClient
+namespace HeimerNET.Lor.GameClient;
+
+/// <summary>
+/// Game screen resolution
+/// </summary>
+public record Screen
 {
     /// <summary>
-    /// Game screen resolution
+    /// Screen width
     /// </summary>
-    public record Screen
-    {
-        /// <summary>
-        /// Screen width
-        /// </summary>
-        [JsonPropertyName("ScreenWidth")]
-        public int Width { get; init; }
+    [JsonPropertyName("ScreenWidth")]
+    public int Width { get; init; }
 
-        /// <summary>
-        /// Screen height
-        /// </summary>
-        [JsonPropertyName("ScreenHeight")]
-        public int Height { get; init; }
-    }
+    /// <summary>
+    /// Screen height
+    /// </summary>
+    [JsonPropertyName("ScreenHeight")]
+    public int Height { get; init; }
 }
