@@ -1,20 +1,19 @@
 ﻿using System.Collections.Generic;
 
-namespace HeimerNET.Lor.GameClient
+namespace HeimerNET.Lor.GameClient;
+
+/// <summary>
+/// Represent the player's deck in the game.
+/// </summary>
+public record StaticDecklist
 {
     /// <summary>
-    /// Represent the player's deck in the game.
+    /// Deckcode of current deck.
     /// </summary>
-    public record StaticDecklist
-    {
-        /// <summary>
-        /// Deckcode of current deck.
-        /// </summary>
-        public string? DeckCode { get; init; }
+    public string? DeckCode { get; init; }
 
-        /// <summary>
-        /// Cards in current deck.
-        /// </summary>
-        public Dictionary<string, int>? CardsInDeck { get; init; }
-    }
+    /// <summary>
+    /// Cards in current deck.
+    /// </summary>
+    public Dictionary<string, int>? CardsInDeck { get; init; }
 }
